Create common module for integration tests
Start moving some of the shared routines used in the integration tests
to a common module. There's still more to move, but this is a good start.
